StorageManagement: Fixed a problem with lazy commits and deletions
Deletions did not return immediately after deleting all rows, and the
deletion had a problem where the number of pages was not decremented on
deletion when the pages are cached and not yet written to disk. So there
were a consistency problem where the number of pages actually is less
than the number reflected in the Table object.

While I'm at it, I also removed a log message that was left
unintentionally.